move trades tables
