+++ packages/ui-components/src/lib/components/detail/TanstackPageContentDetail.svelte
