+++ packages/ui-components/src/lib/index.ts
@@ -27,7 +27,8 @@ export { default as OrderTradesChart } from './components/charts/OrderTradesChar
 export { default as TableTimeFilters } from './components/charts/TableTimeFilters.svelte';
 export { default as OrderTradesListTable } from './components/tables/OrderTradesListTable.svelte';
 export { default as Checkbox } from './components/checkbox/Checkbox.svelte';
-
+export { default as TanstackPageContentDetail } from './components/detail/TanstackPageContentDetail.svelte';
+export { default as EditableSpan } from './components/EditableSpan.svelte';
 //Types
 export type { AppStoresInterface } from './types/appStores.ts';
 export type { ConfigSource, OrderbookConfigSource, OrderbookRef } from './typeshare/config';
